CachingCalculator base class
allows transparent access to calculations that are:

 * calculated on-demand -- if you don't end up needing it, or needing it
   again after an invalidation, it's not (re)calculated
 * cached -- if nothing it depends on changes, it's not recalculated
 * dependency-aware (if something a calculation depends on changes, the
   cache is automatically invalidated).

Simply sub-class CachingCalculator, then put the calculations you want
cached in methods titled 'calculate_X'. an read-only attribute named X
will then be created for transparent access. Feel free to use this
variable in other calculations, which can be used in others, etc.

Construction of the dependency graph is automatic, discovered at runtime
as calculations are performed. (NOTE: this means that if calculate_X
contains conditional logic, dependencies in unexecuted branches will not
be discovered. However, unless something controlling the branching logic
changes, that doesn't matter, and if that something controlling the
branching logic changes, the cache of the calculation will be
invalidated anyways.)

To indicate which non-calculated parameters should be included in the
inferred dependency graph, and thus able to invalidate dependent
calculations, (e.g. level and target_level in DamageCalculator), simply
list their names in the special '__parameters__' class attribute.

The main drawback is that only calculations/parameters defined on the
same class are tracked in the automatic dependency graph. If you need to
invalidate a calculation when a 'foreign' calculation/parameter changes,
you will need to set up a 'native' parameter representing the foreign
value, and then manage that relationship out-of-band.